+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 Features:
 
 - Implement the `IsValidFirestoreId` decorator.
+- Implement the NestJS `FirebaseModule`.
+- Implement the `getDefaultFirebaseApp()` function.